Added unit tests for disease containment and infection

+++ tests/UnitTests/PersonTests.cpp
@@ -6,7 +6,7 @@ namespace UnitTests {
     protected:
         std::string name = "a";
         std::pair<uint16_t, uint16_t> incubationPeriod{ 1, 1 };
-        int daysInfectious = 1;
+        uint16_t daysInfectious = 1;
         std::pair<uint16_t, uint16_t> diseaseDurationRange{ 2, 10 };
         std::vector<float> mortalityByAge{ 1.f, 2.f, 3.f, 4.f, 5.f, 6.f, 7.f, 8.f, 9.f };
         std::pair<uint16_t, uint16_t> daysTillDeathRange{ 1, 2 };
